+++ inversion_count/c/inversions.h
@@ -11,7 +11,7 @@
  */
 typedef int (*comparator)(const void* x, const void* y);
 
-int count_inversions(const void* arr, const size_t length,
+unsigned long count_inversions(const void* arr, const size_t length,
     const size_t item_size, comparator comparator);
 
 #endif
